fullPort and known24s bug fix, disable knownIps

+++ minecraftCheck.js
@@ -7,8 +7,8 @@ function ping(ip, port, protocol, timeout) {
   return new Promise((resolve, reject) => {
     setTimeout(() => {
       if (!hasResponded) {
-        resolve(false);
         client.destroy();
+        resolve(false);
       }
     }, timeout);
     var hasResponded = false;
@@ -19,35 +19,31 @@ function ping(ip, port, protocol, timeout) {
         Buffer.from([0x00]), // packet ID
         Buffer.from(varint.encode(protocol)), //protocol version
         Buffer.from([ip.length]),
-        Buffer.from(ip, 'utf-8'), // server address
-        Buffer.from(new Uint16Array([port]).buffer).reverse(), // server port
+        Buffer.from(ip, 'utf-8'), // ip
+        Buffer.from(new Uint16Array([port]).buffer).reverse(), // port
         Buffer.from([0x01]), // next state (2)
+        Buffer.from([0x01]), // status request size
         Buffer.from([0x00]) // status request
       ]);
       var packetLength = Buffer.alloc(1);
-      packetLength.writeUInt8(handshakePacket.length);
-      var buffer = Buffer.concat([packetLength, handshakePacket]);
+      packetLength.writeUInt8(handshakePacket.length - 2);
+      const buffer = Buffer.concat([packetLength, handshakePacket]);
       client.write(buffer);
     });
 
     client.on('data', (data) => {
-      data = data.toString();
       try {
-        if (JSON.parse(data).players === undefined) console.log(data);
-      } catch (error) {
-        //console.log(data);
-      }
-      client.destroy();
-      resolve(true);
+        varint.decode(data);
+        if (data.slice(varint.decode.bytes)[0] == 0) {
+          client.destroy();
+          resolve(true);
+        }
+      } catch (e) {}
     });
 
-    client.on('error', () => {
-      client.destroy();
-    });
+    client.on('error', client.destroy );
 
-    client.on('close', () => {
-      client.destroy();
-    });
+    client.on('close', client.destroy );
   })
 }
 
@@ -60,12 +56,25 @@ function readIndex(fd, index, size) {
 
 module.exports = (ipsPath, newPath, flag = 'w') => {
   return new Promise(async (resolve, reject) => {
+    const dupeCheck = new Set();
+    const queue = [];
     const writeStream = fs.createWriteStream(newPath, { flags: flag });
     const serverListFD = await (new Promise((resolve, reject) => { fs.open(ipsPath, 'r', (err, fd) => { resolve(fd); }) }));
     const totalServers = fs.statSync(ipsPath).size / 6;
     console.log(`Total servers: ${totalServers}`);
-    const verified = [];
     var serversPinged = 0;
+    var totalResults = 0;
+    var startTime;
+
+    async function write() {
+      if (queue.length > 0) {
+        const buffer = Buffer.concat(queue);
+        queue.splice(0);
+        if (!writeStream.write(buffer)) await new Promise((res) => writeStream.once('drain', res));
+      }
+      setTimeout(write);
+    }
+    write();
 
     async function getServer(i) {
       const server = await readIndex(serverListFD, i * 6, 6);
@@ -76,14 +85,15 @@ module.exports = (ipsPath, newPath, flag = 'w') => {
   
     async function pingServer(serverIndex) {
       serversPinged++;
-      if (serversPinged % 20000 == 0) console.log(serversPinged);
-      if (verified.includes(serverIndex)) return;
+      if (serversPinged % 20000 == 0) console.log(`${serversPinged}/${totalServers} (${Math.floor(serversPinged / totalServers * 100)}%)  Estimated ${Math.floor((totalServers - serversPinged) * (serversPinged / (new Date() - startTime)) / 60000)}:${(totalServers - serversPinged) * (serversPinged / (new Date() - startTime)) % 60000 < 10000 ? 0 : ''}${Math.floor(((totalServers - serversPinged) * (serversPinged / (new Date() - startTime)) % 60000) / 1000)} remaining`);
+      if (dupeCheck.has(serverIndex)) return;
       const server = await getServer(serverIndex);
       try {
         if (await ping(server.ip, server.port, 0, rescanTimeout)) {
-          verified.push(serverIndex);
+          totalResults++;
+          dupeCheck.add(serverIndex);
           const splitIP = server.ip.split('.');
-          writeStream.write(Buffer.from([
+          queue.push(Buffer.from([
             parseInt(splitIP[0]),
             parseInt(splitIP[1]),
             parseInt(splitIP[2]),
@@ -97,34 +107,25 @@ module.exports = (ipsPath, newPath, flag = 'w') => {
   
     function scanBatchPromise(i, startNum) {
       return new Promise((resolve, reject) => {
-        function scanBatch(i, startNum) {
+        async function scanBatch(i, startNum) {
           if (i >= startNum) {
             if (i + rescanRate < totalServers) {
               // scan through the end of the server list
-              for (var j = i; j < i + rescanRate; j++) {
-                pingServer(j)
-              }
-              setTimeout(function() { scanBatch(i + rescanRate, startNum) }, rescanTimeout);
+              for (var j = i; j < i + rescanRate; j++) pingServer(j);
+              setTimeout(() => { scanBatch(i + rescanRate, startNum) }, rescanTimeout);
             } else {
               // once the end of the list is reached, restart at the beginning
-              for (var j = i; j < totalServers; j++) {
-                pingServer(j)
-              }
-              setTimeout(function() { scanBatch(0, startNum) }, rescanTimeout);
+              for (var j = i; j < totalServers; j++) pingServer(j);
+              setTimeout(() => { scanBatch(0, startNum) }, rescanTimeout);
             }
           } else {
             // scan up to the server that was started with (after restarting at the beginning)
             if (i + rescanRate < startNum) {
-              for (var j = i; j < i + rescanRate; j++) {
-                pingServer(j)
-              }
-              setTimeout(function() { scanBatch(i + rescanRate, startNum) }, rescanTimeout);
+              for (var j = i; j < i + rescanRate; j++) pingServer(j);
+              setTimeout(() => { scanBatch(i + rescanRate, startNum) }, rescanTimeout);
             } else {
-              for (var j = i; j < startNum - i; j++) {
-                pingServer(j)
-              }
-      
-              writeStream.close();
+              for (var j = i; j < startNum - i; j++) pingServer(j);
+
               resolve();
             }
           }
@@ -135,12 +136,22 @@ module.exports = (ipsPath, newPath, flag = 'w') => {
 
     for (var i = 0; i < rescans; i++) {
       serversPinged = 0;
+
       var startNum = Math.floor(Math.random() * Math.floor(totalServers / rescanRate)) * rescanRate;
       if (startNum == 0) startNum = rescanRate;
-      const startTime = new Date();
+      startTime = new Date();
       await scanBatchPromise(startNum, startNum);
-      console.log(`Finished scan ${i + 1}/${rescans} in ${(new Date() - startTime) / 1000} seconds at ${new Date().toLocaleString("en-US", { timeZone: "America/Los_Angeles" })}.`);
-      resolve();
+      console.log(`Finished scanning ${totalResults} servers on scan ${i + 1}/${rescans} in ${(new Date() - startTime) / 1000} seconds at ${new Date().toLocaleString()}.`);
     }
+    await (new Promise(res => {
+      const interval = setInterval(() => {
+        if (queue.length == 0) {
+          clearInterval(interval);
+          res();
+        } else console.log(`Finishing write queue: ${queue.length} servers remanining.`);
+      }, 300);
+    }));
+    writeStream.close();
+    resolve();
   })
 }
